+++ test/index.test.js
@@ -1,6 +1,6 @@
 import { assert } from "chai";
 
-describe("Passing test", () => {
+describe("Dummy test", () => {
   it("should pass a dummy test", () => {
     const value = true;
     const expected = true;